uint64_t -> uint32_t since murmurhash returns uint32_t, so uint64_t wouldn't work anyway

+++ src/bloomfilter/bloomfilter.c
@@ -1,8 +1,8 @@
 #include "bloomfilter.h"
 
-Bloomfilter* create_bloomfilter(CategoryList* list, int64_t except, uint8_t p) {
-    uint64_t n8 = list->elements_size * p; // amount of bits (n = m * p)
-    uint64_t n = (n8 + 7) / 8; // amount of bits / 8 rounded up
+Bloomfilter* create_bloomfilter(CategoryList* list, int32_t except, uint8_t p) {
+    uint32_t n8 = list->elements_size * p; // amount of bits (n = m * p)
+    uint32_t n = (n8 + 7) / 8; // amount of bits / 8 rounded up
     uint8_t k = p * log(2); // amount of hashfunctions (k = n/m * ln(2) rounded down = p * ln(2) rounded down)
 
     Bloomfilter* bloomfilter = malloc(sizeof(Bloomfilter));
@@ -17,14 +17,14 @@ Bloomfilter* create_bloomfilter(CategoryList* list, int64_t except, uint8_t p) {
         return clean_return(1, bloomfilter, free_bloomfilter);
     }
 
-    for (int64_t i = 0; i < list->categories_size; i++) {
+    for (int32_t i = 0; i < list->categories_size; i++) {
         if (i == except) continue;
         Category* category = list->categories[i];
-        for (int64_t j = 0; j < category->size; j++) {
+        for (int32_t j = 0; j < category->size; j++) {
             char* element = category->elements[j];
             uint8_t element_length = strlen(element);
             for (int8_t l = 0; l < k; l++) {
-                uint64_t hash = murmurhash(element, element_length, l) % (n * 8);
+                uint32_t hash = murmurhash(element, element_length, l) % (n * 8);
                 bloomfilter->bf[hash / 8] |= (1ULL << (hash % 8));
             }
         }
